form updates user and messages

+++ flask/Unit-01/06-sql-alchemy-2/app.py
@@ -2,11 +2,14 @@ from flask import Flask, request, redirect, url_for, render_template
 from flask_modus import Modus
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
+from forms import UserForm, MessageForm, DeleteForm
+import os
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'postgres://localhost/troll'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 app.config['SQLALCHEMY_ECHO'] = True
+app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY')
 modus = Modus(app)
 db = SQLAlchemy(app)
 Migrate(app, db)
@@ -50,47 +53,63 @@ def index():
 
 @app.route('/users/new')
 def new():
-    return render_template('user/new.html')
+    user_form = UserForm()
+    return render_template('user/new.html', form=user_form)
 
 
 @app.route('/users', methods=['POST'])
 def create():
-    user = User(
-        first_name=request.form.get('first_name'),
-        last_name=request.form.get('last_name'))
-    db.session.add(user)
-    db.session.commit()
-    return redirect(url_for('index'))
+    user_form = UserForm(request.form)
+    #if form is validate commit if not take info and render template
+    if user_form.validate():
+        user = User(
+            first_name=request.form.get('first_name'),
+            last_name=request.form.get('last_name'))
+        db.session.add(user)
+        db.session.commit()
+        return redirect(url_for('index'))
+    else:
+        return render_template('user/new.html', form=user_form)
 
 
 @app.route('/users/<int:id>')
 def show(id):
+    delete_form = DeleteForm()
     found_user = User.query.get(id)
-    return render_template('user/show.html', user=found_user)
+    return render_template('user/show.html', user=found_user, delete_form=delete_form)
 
 
 @app.route('/users/<int:id>/edit')
 def edit(id):
     found_user = User.query.get(id)
-    return render_template('user/edit.html', user=found_user)
+    #obj=found_user prepopulates the edit form
+    user_form = UserForm(obj=found_user)
+    return render_template('user/edit.html', user=found_user, form=user_form)
 
 
 @app.route('/users/<int:id>', methods=['PATCH'])
 def update(id):
-    found_user = User.query.get(id)
-    found_user.first_name = request.form.get('first_name')
-    found_user.last_name = request.form.get('last_name')
-    db.session.add(found_user)
-    db.session.commit()
-    return redirect(url_for('index'))
-
+    user_form = UserForm(request.form)
+    if user_form.validate():
+        found_user = User.query.get(id)
+        found_user.first_name = request.form.get('first_name')
+        found_user.last_name = request.form.get('last_name')
+        db.session.add(found_user)
+        db.session.commit()
+        return redirect(url_for('index'))
+    else:
+        return render_template('user/edit.html', user=found_user, form=user_form)
 
 @app.route('/users/<int:id>', methods=['DELETE'])
 def destroy(id):
-    found_user = User.query.get(id)
-    db.session.delete(found_user)
-    db.session.commit()
-    return redirect(url_for('index'))
+    delete_form = DeleteForm(request.form)
+    if delete_form.validate():
+        found_user = User.query.get(id)
+        db.session.delete(found_user)
+        db.session.commit()
+        return redirect(url_for('index'))
+    else:
+        return render_template('user/show.html', user=found_user, delete_form=delete_form)
 
 
 # ROUTES FOR MESSAGES NESTED INSIDE USER
@@ -110,7 +129,8 @@ def messages_index(user_id):
 
 @app.route('/users/<int:user_id>/messages/new')
 def messages_new(user_id):
-    return render_template('message/new.html', user=User.query.get(user_id))
+    message_form = MessageForm()
+    return render_template('message/new.html', user=User.query.get(user_id), form=message_form)
 
 
 #create a specific message for a specfic user
@@ -118,20 +138,27 @@ def messages_new(user_id):
 
 @app.route('/users/<int:user_id>/messages', methods=["POST"])
 def messages_create(user_id):
-    new_message = Message(message=request.form.get('message'), user_id=user_id)
-    db.session.add(new_message)
-    db.session.commit()
-    return redirect(url_for('messages_index', user_id=user_id))
+    message_form = MessageForm(request.form)
+    if message_form.validate():
+        new_message = Message(message=request.form.get('message'), user_id=user_id)
+        db.session.add(new_message)
+        db.session.commit()
+        return redirect(url_for('messages_index', user_id=user_id))
+    else:
+        return render_template('message/new.html', user=User.query.get(user_id), form=message_form)
 
 
 #edit a specific message for a specfic user render template
 @app.route('/users/<int:user_id>/messages/<int:message_id>/edit')
 def messages_edit(user_id, message_id):
+    found_message = Message.query.get(message_id)
+    found_user = User.query.get(user_id)
+    message_form = MessageForm(obj=found_message)
     return render_template(
         'message/edit.html',
         #don't name varibales same name as parameters dumbo Mark
-        user=User.query.get(user_id),
-        message=Message.query.get(message_id))
+        user=found_user,
+        message=found_message, form=message_form)
 
 
 #show message
@@ -142,11 +169,18 @@ def messages_show(user_id, message_id):
 #update message on submit after edit
 @app.route('/users/<int:user_id>/messages/<int:message_id>', methods=["PATCH"])
 def messages_update(user_id, message_id):
-    update_message = Message.query.get(message_id)
-    update_message.message = request.form.get('message')
-    db.session.add(update_message)
-    db.session.commit()
-    return redirect(url_for('messages_index', user_id=user_id))
+    message_form=MessageForm(request.form)
+    #NEED TO PASS IN FOR THE REDIRECT
+    found_user = User.query.get(user_id)
+    found_message = Message.query.get(message_id)
+    if message_form.validate():
+        update_message = Message.query.get(message_id)
+        update_message.message = request.form.get('message')
+        db.session.add(update_message)
+        db.session.commit()
+        return redirect(url_for('messages_index', user_id=user_id))
+    else:
+        return render_template('message/edit.html', user=found_user, message=found_message, form=message_form)
 
 #delete a specific message for a specfic user
 @app.route('/users/<int:user_id>/messages/<int:message_id>', methods=["DELETE"])
@@ -159,7 +193,8 @@ def messages_destroy(user_id, message_id):
 
 @app.errorhandler(404)
 def page_not_found(e):
-    return render_template('404.html',404)
+    return render_template('user/404.html'), 404
+
 # Mistakes
 # dont name variables the same name as parameters or things will be undefined
 # dont forget redirect or render_template after return
